# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ContextForge API
  description: |
    ContextForge API provides comprehensive context management capabilities with AI-powered features.
    
    ## Authentication
    
    The API supports multiple authentication methods:
    - API Key authentication via `Authorization: Bearer <token>` header
    - Session-based authentication via cookies
    
    ## Rate Limiting
    
    API requests are rate limited based on your subscription plan:
    - Free: 100 requests/hour
    - Pro: 1,000 requests/hour  
    - Enterprise: 10,000 requests/hour
    
    ## Error Handling
    
    The API uses conventional HTTP response codes:
    - 2xx: Success
    - 4xx: Client errors
    - 5xx: Server errors
    
    All errors return a JSON object with an `error` field describing the issue.
    
  version: 1.0.0
  contact:
    name: ContextForge Support
    url: https://contextforge.com/support
    email: support@contextforge.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.contextforge.com/v1
    description: Production server
  - url: https://staging-api.contextforge.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
security:
  - BearerAuth: []
  - CookieAuth: []
...
paths:
...
  /intelligence/search:
    post:
      tags: [Intelligence]
      summary: Semantic search
      description: Perform AI-powered semantic search across content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  description: Search query
                  example: "How to optimize database queries"
                types:
                  type: array
                  items:
                    type: string
                    enum: [prompt, rule, agent, collection]
                limit:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 10
                threshold:
                  type: number
                  minimum: 0
                  maximum: 1
                  default: 0.5
                  description: Similarity threshold
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/ContextItem'
                        - type: object
                          properties:
                            similarity:
                              type: number
                              description: Similarity score
                  query:
                    type: string
                  total:
                    type: integer
  /intelligence/optimization:
    post:
      tags: [Intelligence]
      summary: Optimize content
      description: AI-powered content optimization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemId]
              properties:
                itemId:
                  type: string
                model:
                  type: string
                  enum: [gpt-4, claude, gemini]
                  default: gpt-4
                creativity:
                  type: number
                  minimum: 0
                  maximum: 1
                  default: 0.7
                focusAreas:
                  type: array
                  items:
                    type: string
                    enum: [clarity, structure, engagement, conciseness]
                  default: [clarity, structure]
                preserveStructure:
                  type: boolean
                  default: false
                preview:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Optimization completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  optimizedContent:
                    type: string
                  improvement:
                    type: number
                    description: Quality improvement percentage
                  changes:
                    type: array
                    items:
                      type: string
                  suggestions:
                    type: array
                    items:
                      type: string
  /intelligence/batch:
    post:
      tags: [Intelligence]
      summary: Batch AI operations
      description: Perform AI operations on multiple items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemIds, operation]
              properties:
                itemIds:
                  type: array
                  items:
                    type: string
                  maxItems: 100
                operation:
                  type: string
                  enum: [optimize, classify, analyze]
                options:
                  type: object
                  description: Operation-specific options
      responses:
        '202':
          description: Batch operation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                  status:
                    type: string
                    enum: [pending, running, completed, failed]
                  estimatedDuration:
                    type: integer
                    description: Estimated duration in seconds

  # Jobs
  /jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get job status
      description: Retrieve status and results of a background job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'

  # Health Check
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth-token
  schemas:
...
    ContextItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [prompt, rule, agent, collection]
        description:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        folderId:
          type: string
        folderPath:
          type: string
        metadata:
          type: object
        quality:
          type: object
          properties:
            score:
              type: number
              minimum: 0
              maximum: 1
            issues:
              type: array
              items:
                type: string
            suggestions:
              type: array
              items:
                type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Job:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [import, export, optimize, classify, analyze]
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress:
          type: number
          minimum: 0
          maximum: 100
        totalItems:
          type: integer
        processedItems:
          type: integer
        results:
          type: object
        error:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...